refactor: move all state handling inside DataList.js chore: update JSDocs
