+++ components/UI/InviteUsersToMatrixRoom/index.js
@@ -1,13 +1,12 @@
 /**
- * This component provides a user interface for inviting other users to a Matrix room.
- * It includes a form for searching and inviting users, and provides feedback to the user.
+ * Provides a user interface for inviting other users to a Matrix room.
  *
- * @param {string} roomId - The valid Matrix room ID to which you want to invite a user.
+ * @param {string} roomId - The ID of the Matrix room to invite users to.
  * @param {string} roomName - The name of the Matrix room.
- * @param {Function} onSuccess - An optional callback function to execute after a successful invitation.
-
- * @returns {React.ReactElement} - A React element representing the component.
-
+ * @param {Function} onSuccess - An optional callback executed after a successful invitation.
+ *
+ * @returns {React.ReactElement} - A React component representing the invitation UI.
+ *
  * @example
  * // Example usage of the InviteUserToMatrixRoom component:
  * <InviteUserToMatrixRoom roomId="your-room-id" roomName="Your Room" onSuccess={handleSuccess} />
@@ -18,14 +17,10 @@ import { useTranslation } from 'react-i18next';
 import { debounce } from 'lodash';
 import { logger } from 'matrix-js-sdk/lib/logger';
 import styled from 'styled-components';
-import { CloseIcon } from 'next/dist/client/components/react-dev-overlay/internal/icons/CloseIcon';
 
-import Form from '../Form';
 import { useAuth } from '../../../lib/Auth';
 import ErrorMessage from '../ErrorMessage';
 import Datalist from '../DataList';
-import { ServiceTable } from '../ServiceTable';
-import TextButton from '../TextButton';
 
 const ActionWrapper = styled.section`
   display: grid;
@@ -37,15 +32,6 @@ const ActionWrapper = styled.section`
 
   h3 {
     line-height: calc(var(--margin) *3);
-  }
-`;
-
-const InviteUserForm = styled(Form)`
-  display: grid;
-  height: 100%;
-
-  > :last-child {
-    align-self: end;
   }
 `;
 
@@ -100,60 +86,29 @@ export default function InviteUserToMatrixRoom({ roomId, onSuccess }) {
 
         const successAmount = selectedUsers.length - errors.length;
 
-        // if everything is okay, we let the user know and exit the modal view.
+        // if everything is okay, we let the user know and exit the view.
         setUserFeedback('✓ ' + successAmount + ' ' + t('{{user}} invited and needs to accept your invitation', { user: successAmount > 1 ? 'users were' : 'user was' }));
         await new Promise(() => setTimeout(() => {
             clearInputs();
-            onSuccess && successAmount === selectedUsers.length && onSuccess();
+            if (onSuccess && successAmount === selectedUsers.length) onSuccess();
         }, 3000));
-    };
 
-    const handleUserSelect = (user) => {
-        const find = selectedUsers.some(selectedUser => selectedUser?.user_id === user.user_id);
-        if (find) setSelectedUsers(prevState => prevState.filter(state => state !== user));
-        else setSelectedUsers(prevState => [...prevState, user]);
-    };
-
-    const handleRemove = (user) => {
-        const filteredArray = selectedUsers.filter(selectedUser => selectedUser?.user_id !== user.user_id);
-        setSelectedUsers(filteredArray);
+        return successAmount === selectedUsers.length;
     };
 
     return <ActionWrapper>
         <h3>{ t('Invite users') }</h3>
-        <InviteUserForm onSubmit={handleInvite}>
-            { userFeedback && !errorFeedback ? <div>{ userFeedback }</div> :
-                <>
-                    <Datalist
-                        options={searchResults}
-                        onChange={handleChange}
-                        keysToDisplay={['display_name', 'user_id']}
-                        selected={selectedUsers}
-                        onSelect={handleUserSelect}
-                    />
-                    { /*{ selectedUsers.length !== 0 && <ServiceTable>{ selectedUsers.map(user => {*/ }
-                    { /*    return <DisplaySelectedUser key={user.display_name} user={user} handleRemove={handleRemove} />;*/ }
-                    { /*},*/ }
-                    { /*) }*/ }
-                    { /*</ServiceTable> }*/ }
-                    <button disabled={selectedUsers.length === 0}>{ t('invite') }</button>
-                </>
-            }
-            { userFeedback && errorFeedback && userFeedback }
-            { errorFeedback && errorFeedback }
-        </InviteUserForm>
+        { userFeedback && !errorFeedback ? <div>{ userFeedback }</div> :
+            <Datalist
+                options={searchResults}
+                onInputChange={handleChange}
+                keysToDisplay={['display_name', 'user_id']}
+                selected={selectedUsers}
+                onSubmit={handleInvite}
+            />
+        }
+        { userFeedback && errorFeedback && userFeedback }
+        { errorFeedback && errorFeedback }
+
     </ActionWrapper>;
 }
-
-const DisplaySelectedUser = ({ user, handleRemove }) => {
-    return <ServiceTable.Row>
-        <ServiceTable.Cell>
-            { user.display_name }
-        </ServiceTable.Cell>
-        <ServiceTable.Cell align="right">
-            <TextButton onClick={() => handleRemove(user)}>
-                <CloseIcon width="var(--icon-size)" height="var(--icon-size)" fill="var(--color-foreground)" />
-            </TextButton>
-        </ServiceTable.Cell>
-    </ServiceTable.Row>;
-};
